- Set up mLab mongoDB - Extract config_server.js and config_client.js
