+++ .gitignore
@@ -1,3 +1,5 @@
+config_local.js
+
 # Logs
 logs
 *.log
