+++ src/server/server.js
@@ -3,10 +3,12 @@ const express = require('express');
 const compression = require('compression');
 const cookieParser = require('cookie-parser');
 
-const { env, port, publicDir } = require('../build/config');
+const { env, port, publicDir } = require('../build/config_server');
 const middlewares = require('./middlewares');
 const CacheProvider = require('./cacheProvider');
 const stockProvider = require('./stockProvider')({ env, axios });
+
+const Mongo = require('./db/mongo'); // eslint-disable-line no-unused-vars
 
 const PUBLIC_DIR = publicDir;
 const PORT = port;
